[plugin.video.plutotv] 1.0.7
